Split store module into submodules
