+++ packages/graph/hash_graph/lib/graph/src/lib.rs
@@ -65,6 +65,8 @@
     reason = "This encourages importing `as` which breaks IDEs"
 )]
 
+extern crate core;
+
 pub mod api;
 
 pub mod knowledge;
